// FashionStylist/angularapp/src/app/model/stylist.model.ts
// export interface Doctor {
//     id?: number;
//     name: string;
//     age: number;
//     specialization: string;
//     department: string;
//     contactNumber: string;
//   }

export interface Stylist {
  id?: number;
  name: string;
  expertise: string; 
  styleSignature: string; 
  availability: string; 
  hourlyRate: number;
  location: string; 
  }[];